Career/career_guidance_api_django/requirements.txt
Django
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
requests>=2.32.0
